requirements.txt
transformers
torch
numpy
Pillow
opencv-python
streamlit
timm